cleanup: strongly types enums

+++ motor_code.cpp
@@ -39,7 +39,7 @@ void Motor::begin_move_with_direction(Motor::Direction n_direction) {
 
   auto motor_speed = static_cast<bool>(this->_current_direction) ? Motor::Max_Speed::Forwards : Motor::Max_Speed::Backwards;
 
-  this->_motor_393.write(motor_speed);
+  this->_motor_393.write(static_cast<int16_t>(motor_speed));
   this->_active = true;
 }
